+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+0.3.0 (2020-01-22)
+------------------
+
+* Dropped support for Python 3.4.
+* ``subtests`` now correctly captures and displays stdout/stderr (`#18`_).
+
+.. _#18: https://github.com/pytest-dev/pytest-subtests/issues/18
+
 0.2.1 (2019-04-04)
 ------------------
 
@@ -11,7 +19,7 @@ CHANGELOG
 0.2.0 (2019-04-03)
 ------------------
 
-* Sub tests are correctly reported with ``pytest-xdist>=1.28``.
+* Subtests are correctly reported with ``pytest-xdist>=1.28``.
 
 0.1.0 (2019-04-01)
 ------------------